split off code into vector_simd.rs
